+++ Assets/Scripe/Bean/VocationDecBean.cs
@@ -16,7 +16,7 @@ public class VocationDecBean
     public string attributeName;
     public string tip_dec;
     public long resource;
-    public float attack_range;
+    public double attack_range;
 
     private List<long> mNexts;
     private List<long> mAttribute;
