+++ Assets/Scripe/SkillPakeger/CardSkill/pointer/SkillTargetEllipseDeal.cs
@@ -19,7 +19,7 @@ public class SkillTargetEllipseDeal
             {
 
                 float x = tmp.mCurrentX+tmp.mAttacker.resourceData.getHurtOffset().x - local.x;
-                float y = tmp.mCurrentY+tmp.mAttacker.resourceData.idel_y - local.y;
+                float y = tmp.mCurrentY+ (float)tmp.mAttacker.resourceData.idel_y - local.y;
                 if (x * x / a + y * y / b <= 1)
                 {
                     result.Add(tmp.mAttacker);
